feat(ssif): handle zero-lengh response for block read as an error
When the upper-layer application on the bmc does not return response
data in time for an ipmi request from the host, the ssif driver on the
bmc side may return 0.

This situation may occur during the ipmi-ssif driver probe and cause
the probe to fail. We should handle this situation as an error and retry.

Signed-off-by: leon_fang <leon_fang@realsil.com.cn>